requirements.txt
selenium>=4.15.0
pytest>=7.4.0
pytest-html>=4.1.0
pytest-xdist>=3.3.0
webdriver-manager>=4.0.0
allure-pytest>=2.13.0
pytest-rerunfailures>=12.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0